Further improve the guards behavior.
Stick to guard/patrol/follow target if attack target is too far away.

+++ src/main/java/com/minecolonies/entity/EntityCitizen.java
@@ -546,14 +546,14 @@ public class EntityCitizen extends EntityAgeable implements INpc
                 LanguageHandler.sendPlayersLocalizedMessage(
                   colony.getMessageEntityPlayers(),
                   "tile.blockHutTownHall.messageGuardDead",
-                  citizenData.getName(), posX, posY, posZ);
+                  citizenData.getName(), (int)posX, (int)posY, (int)posZ);
             }
             else
             {
                 LanguageHandler.sendPlayersLocalizedMessage(
                   colony.getMessageEntityPlayers(),
                   "tile.blockHutTownHall.messageColonistDead",
-                  citizenData.getName(), posX, posY, posZ);
+                  citizenData.getName(), (int) posX, (int)posY, (int)posZ);
             }
             colony.removeCitizen(getCitizenData());
         }
